update: add stack chapter's multiple solutions

+++ Part2_강의자료/Ch02_스택/문제별코드/10799_쇠막대기/10799.java
@@ -7,7 +7,7 @@ class Main
 
         char[] input = sc.next().toCharArray();
         int openCount = 0;
-        long ans = 0;
+        int ans = 0;
         for (int i = 0; i < input.length; i++) {
             if (input[i] == '(') openCount++;
             else {
